Generate main toc by collecting headers
Before, doctoc collected the transformAndSave output to make the final toc. Now it collects header objects instead, which are passed to transformAndSave to make the main toc. The advantage is that the toc generation code can stay inside transform.js instead of being split between transform.js and doctoc.js.